restructure layout folder and update styles to match Figma
