+++ ui/src/App.tsx
@@ -1,55 +1,31 @@
-// Copyright 2024 Iguazio
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+Copyright 2024 Iguazio Systems Ltd.
 
-import { ChatPage } from 'pages/ChatPage';
-import { LoginPage } from 'pages/LoginPage';
-import { ProjectPage } from 'pages/ProjectPage';
-import { ProjectsPage } from 'pages/ProjectsPage';
-import { UsersTablePage } from 'pages/UsersTablePage';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+Licensed under the Apache License, Version 2.0 (the "License") with
+an addition restriction as set forth herein. You may not use this
+file except in compliance with the License. You may obtain a copy of
+the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+In addition, you may not use the software for any purposes that are
+illegal under applicable law, and the grant of the foregoing license
+under the Apache 2.0 license is conditioned upon your compliance with
+such restriction.
+*/
+
+import { RouterProvider } from 'react-router-dom';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
+import { router } from './router';
+
 function App() {
   const queryClient = new QueryClient();
-
-  const router = createBrowserRouter([
-    {
-      path: '/',
-      element: <LoginPage />,
-    },
-    {
-      path: '/projects',
-      element: <ProjectsPage />,
-    },
-    {
-      path: '/projects/:name',
-      element: <ProjectPage />,
-    },
-    {
-      path: '/users',
-      element: <UsersTablePage />,
-    },
-    {
-      path: '/chat',
-      element: <ChatPage />,
-    },
-    {
-      path: '/chat/:sessionName',
-      element: <ChatPage />,
-    },
-  ]);
 
   return (
     <QueryClientProvider client={queryClient}>
