+++ ui/src/layout/Navbar/Navbar.test.tsx
@@ -1,16 +1,22 @@
-// Copyright 2024 Iguazio
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+Copyright 2024 Iguazio Systems Ltd.
+
+Licensed under the Apache License, Version 2.0 (the "License") with
+an addition restriction as set forth herein. You may not use this
+file except in compliance with the License. You may obtain a copy of
+the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+In addition, you may not use the software for any purposes that are
+illegal under applicable law, and the grant of the foregoing license
+under the Apache 2.0 license is conditioned upon your compliance with
+such restriction.
+*/
 
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
@@ -20,7 +26,7 @@ import theme from '@shared/theme';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { fireEvent, render, screen } from '@testing-library/react';
 
-import Topbar from './Topbar';
+import Navbar from './Navbar';
 
 const renderWithProviders = (ui: React.ReactElement) => {
   const queryClient = new QueryClient();
@@ -55,21 +61,19 @@ window.matchMedia =
   }));
 
 describe('Topbar Component (Zustand)', () => {
-  const mockOnLoginChange = jest.fn();
-
   beforeEach(() => {
     jest.clearAllMocks();
   });
 
   it('renders Topbar component with logo and avatar', () => {
-    renderWithProviders(<Topbar onLoginChange={mockOnLoginChange} />);
+    renderWithProviders(<Navbar />);
 
     expect(screen.getByTestId('logo')).toBeInTheDocument();
     expect(screen.getByTestId('avatar')).toBeInTheDocument();
   });
 
   it('opens Rightbar when avatar is clicked', () => {
-    renderWithProviders(<Topbar onLoginChange={mockOnLoginChange} />);
+    renderWithProviders(<Navbar />);
 
     const avatar = screen.getByTestId('avatar');
     fireEvent.click(avatar);
