+++ ui/src/pages/ProjectsPage.tsx
@@ -1,16 +1,22 @@
-// Copyright 2024 Iguazio
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+Copyright 2024 Iguazio Systems Ltd.
+
+Licensed under the Apache License, Version 2.0 (the "License") with
+an addition restriction as set forth herein. You may not use this
+file except in compliance with the License. You may obtain a copy of
+the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+In addition, you may not use the software for any purposes that are
+illegal under applicable law, and the grant of the foregoing license
+under the Apache 2.0 license is conditioned upon your compliance with
+such restriction.
+*/
 
 import { useNavigate } from 'react-router-dom';
 
@@ -25,7 +31,6 @@ import {
   useToast,
   VStack,
 } from '@chakra-ui/react';
-import Layout from '@components/feature/Layout';
 import AddEditProjectModal from '@components/feature/Projects/AddEditProjectModal';
 import ProjectCard from '@components/feature/Projects/ProjectCard';
 import Loading from '@components/shared/Loading';
@@ -68,51 +73,49 @@ export const ProjectsPage = () => {
   if (isLoading) return <Loading />;
 
   return (
-    <Layout>
-      <Container maxW="container.lg" py={8}>
-        <VStack spacing={6} align="start">
-          <Flex width="100%" justifyContent={'space-between'}>
-            <Heading as="h2" size="lg">
-              Projects
-            </Heading>
-            <Flex gap={4}>
-              <Button
-                leftIcon={<AddIcon />}
-                onClick={() => {
-                  onModalOpen();
-                }}
-              >
-                New
-              </Button>
-              <Button
-                colorScheme="blue"
-                alignSelf="end"
-                onClick={handleOpenPlayground}
-              >
-                Open Playground
-              </Button>
-            </Flex>
+    <Container maxW="container.lg" py={8}>
+      <VStack spacing={6} align="start">
+        <Flex width="100%" justifyContent={'space-between'}>
+          <Heading as="h2" size="lg">
+            Projects
+          </Heading>
+          <Flex gap={4}>
+            <Button
+              leftIcon={<AddIcon />}
+              onClick={() => {
+                onModalOpen();
+              }}
+            >
+              New
+            </Button>
+            <Button
+              colorScheme="blue"
+              alignSelf="end"
+              onClick={handleOpenPlayground}
+            >
+              Open Playground
+            </Button>
           </Flex>
+        </Flex>
 
-          <SimpleGrid columns={[1, null, 2]} spacing={6} w="100%">
-            {projects ? (
-              projects.map((project) => (
-                <ProjectCard
-                  key={project.uid ?? project.name}
-                  project={project}
-                />
-              ))
-            ) : (
-              <div>No projects</div>
-            )}
-          </SimpleGrid>
-        </VStack>
-      </Container>
+        <SimpleGrid columns={[1, null, 2]} spacing={6} w="100%">
+          {projects ? (
+            projects.map((project) => (
+              <ProjectCard
+                key={project.uid ?? project.name}
+                project={project}
+              />
+            ))
+          ) : (
+            <div>No projects</div>
+          )}
+        </SimpleGrid>
+      </VStack>
       <AddEditProjectModal
         isOpen={isModalOpen}
         onClose={onModalClose}
         onSave={handleSave}
       />
-    </Layout>
+    </Container>
   );
 };
