+++ ui/src/router.tsx
@@ -1,0 +1,52 @@
+/*
+Copyright 2024 Iguazio Systems Ltd.
+
+Licensed under the Apache License, Version 2.0 (the "License") with
+an addition restriction as set forth herein. You may not use this
+file except in compliance with the License. You may obtain a copy of
+the License at http://www.apache.org/licenses/LICENSE-2.0.
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+implied. See the License for the specific language governing
+permissions and limitations under the License.
+
+In addition, you may not use the software for any purposes that are
+illegal under applicable law, and the grant of the foregoing license
+under the Apache 2.0 license is conditioned upon your compliance with
+such restriction.
+*/
+
+import { ChatPage } from 'pages/ChatPage';
+import { LoginPage } from 'pages/LoginPage';
+import { ProjectPage } from 'pages/ProjectPage';
+import { ProjectsPage } from 'pages/ProjectsPage';
+import { UsersTablePage } from 'pages/UsersTablePage';
+import { createBrowserRouter, Navigate } from 'react-router-dom';
+
+import Page from '@layout/Page';
+
+export const router = createBrowserRouter([
+  {
+    path: '/',
+    element:
+      import.meta.env.VITE_ENV === 'development' ? (
+        // TODO: Will be implemented GAIT-36
+        <Navigate to="/users" replace />
+      ) : (
+        <LoginPage />
+      ),
+  },
+  {
+    path: '/',
+    element: <Page />,
+    children: [
+      { path: 'projects', element: <ProjectsPage /> },
+      { path: 'projects/:name', element: <ProjectPage /> },
+      { path: 'users', element: <UsersTablePage /> },
+      { path: 'chat', element: <ChatPage /> },
+      { path: 'chat/:sessionName', element: <ChatPage /> },
+    ],
+  },
+]);
